Move http code to http package
