+++ internal/http/error.go
@@ -13,3 +13,29 @@ func (e invalidRequestError) Error() string {
 func (e invalidRequestError) ClientError() string {
 	return "invalid_request"
 }
+
+// noStatusCodeError helps the errorHandler do decide, if an status code can be
+// set.
+type noStatusCodeError struct {
+	wrapped error
+}
+
+func (e noStatusCodeError) Error() string {
+	return e.wrapped.Error()
+}
+
+func (e noStatusCodeError) Unwrap() error {
+	return e.wrapped
+}
+
+func (e noStatusCodeError) NoStatus() {}
+
+type authRequiredError struct{}
+
+func (e authRequiredError) Error() string {
+	return "Only authenticated users can use the notify system."
+}
+
+func (e authRequiredError) ClientError() string {
+	return "auth_required"
+}
