+++ internal/http/http.go
@@ -2,11 +2,17 @@ package http
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
 	"log"
 	"net/http"
+	"strconv"
+	"strings"
+	"sync"
 
+	"github.com/OpenSlides/openslides3-autoupdate-service/internal/auth"
 	"github.com/OpenSlides/openslides3-autoupdate-service/internal/autoupdate"
 	"github.com/OpenSlides/openslides3-autoupdate-service/internal/notify"
 )
@@ -30,28 +36,157 @@ func Health(mux *http.ServeMux) {
 }
 
 // Autoupdate registers the autoupdate route.
-func Autoupdate(mux *http.ServeMux, auto *autoupdate.Autoupdate, auth Auther) {
-	mux.Handle("/system/autoupdate", errHandleFunc(middleware(auto.HandleAutoupdate, auth)))
+func Autoupdate(mux *http.ServeMux, auto *autoupdate.Autoupdate, auther Auther) {
+	handler := func(w http.ResponseWriter, r *http.Request) error {
+		w.Header().Set("Content-Type", "application/octet-stream")
+
+		rawChangeID := r.URL.Query().Get("change_id")
+		var changeID int
+		if rawChangeID != "" {
+			var err error
+			changeID, err = strconv.Atoi(rawChangeID)
+			if err != nil {
+				return invalidRequestError{fmt.Errorf("Change id has to be a number not %s", rawChangeID)}
+			}
+		}
+
+		w.WriteHeader(http.StatusOK)
+		w.(http.Flusher).Flush()
+
+		// Retrive uid from request. 0 for anonymous.
+		uid := auth.FromContext(r.Context())
+		log.Printf("connect user %d with change_id %d", uid, changeID)
+
+		for {
+			all, data, newChangeID, err := auto.Receive(r.Context(), uid, changeID)
+			if err != nil {
+				return noStatusCodeError{err}
+			}
+
+			if len(data) == 0 {
+				continue
+			}
+
+			if err := sendAutoupdateData(w, all, data, changeID, newChangeID); err != nil {
+				return noStatusCodeError{err}
+			}
+			changeID = newChangeID
+		}
+	}
+
+	mux.Handle("/system/autoupdate", errHandleFunc(middleware(handler, auther)))
 }
 
 // Projector registers the projector route.
 func Projector(mux *http.ServeMux, auto *autoupdate.Autoupdate, auth Auther) {
-	mux.Handle("/system/projector", errHandleFunc(middleware(auto.HandleProjector, auth)))
+	count := new(connectionCount)
+
+	handler := func(w http.ResponseWriter, r *http.Request) error {
+		n := count.Add()
+		log.Println("Got projector connection. Connection count: ", n)
+
+		defer func() {
+			n := count.Sub()
+			log.Println("Lost projector connection. Connection count: ", n)
+		}()
+
+		w.Header().Set("Content-Type", "application/json")
+
+		projectorIDs, err := projectorIDs(r.URL.Query().Get("projector_ids"))
+		if err != nil {
+			return err
+		}
+
+		w.WriteHeader(http.StatusOK)
+		w.(http.Flusher).Flush()
+
+		encoder := json.NewEncoder(w)
+		var tid uint64
+		for {
+			ntid, data, cid, err := auto.Projectors(r.Context(), tid, projectorIDs)
+			if err != nil {
+				return noStatusCodeError{err}
+			}
+
+			out := struct {
+				CID  int                     `json:"change_id"`
+				Data map[int]json.RawMessage `json:"data"`
+			}{
+				cid,
+				data,
+			}
+
+			if err := encoder.Encode(out); err != nil {
+				return noStatusCodeError{err}
+			}
+			w.(http.Flusher).Flush()
+			tid = ntid
+		}
+	}
+	mux.Handle("/system/projector", errHandleFunc(middleware(handler, auth)))
 }
 
 // Notify registers the notify route.
-func Notify(mux *http.ServeMux, n *notify.Notify, auth Auther) {
-	mux.Handle("/system/notify", errHandleFunc(middleware(n.HandleNotify, auth)))
+func Notify(mux *http.ServeMux, n *notify.Notify, auther Auther) {
+	handler := func(w http.ResponseWriter, r *http.Request) error {
+		w.Header().Set("Content-Type", "application/octet-stream")
+
+		userID := auth.FromContext(r.Context())
+		if userID == 0 {
+			return authRequiredError{}
+		}
+
+		cid := n.GenerateChannelID(userID)
+		tid := n.LastID()
+
+		w.WriteHeader(http.StatusOK)
+
+		if _, err := fmt.Fprintf(w, `{"channel_id": "%s"}`, cid); err != nil {
+			return noStatusCodeError{err}
+		}
+
+		if _, err := fmt.Fprintln(w); err != nil {
+			return noStatusCodeError{err}
+		}
+		w.(http.Flusher).Flush()
+
+		encoder := json.NewEncoder(w)
+		var err error
+
+		for {
+			tid, err = n.Receive(r.Context(), w, tid, userID, cid, encoder)
+			if err != nil {
+				return noStatusCodeError{err}
+			}
+			w.(http.Flusher).Flush()
+		}
+	}
+	mux.Handle("/system/notify", errHandleFunc(middleware(handler, auther)))
 }
 
 // NotifySend registers the notify/send route.
-func NotifySend(mux *http.ServeMux, n *notify.Notify, auth Auther) {
-	mux.Handle("/system/notify/send", errHandleFunc(middleware(n.HandleSend, auth)))
+func NotifySend(mux *http.ServeMux, n *notify.Notify, auther Auther) {
+	handler := func(w http.ResponseWriter, r *http.Request) error {
+		userID := auth.FromContext(r.Context())
+		if userID == 0 {
+			return authRequiredError{}
+		}
+		return n.Send(r.Body, userID)
+	}
+	mux.Handle("/system/notify/send", errHandleFunc(middleware(handler, auther)))
 }
 
 // NotifyApplause registers the notify/applause route.
-func NotifyApplause(mux *http.ServeMux, n *notify.Notify, auth Auther) {
-	mux.Handle("/system/applause", errHandleFunc(middleware(n.HandleApplause, auth)))
+func NotifyApplause(mux *http.ServeMux, n *notify.Notify, auther Auther) {
+	handler := func(w http.ResponseWriter, r *http.Request) error {
+		userID := auth.FromContext(r.Context())
+		if userID == 0 {
+			return authRequiredError{}
+		}
+
+		return n.AddApplause(userID)
+	}
+	mux.Handle("/system/applause", errHandleFunc(middleware(handler, auther)))
 }
 
 // errHandleFunc is like a http.Handler, but has a error as return value.
@@ -137,4 +272,85 @@ func middleware(next errHandleFunc, a Auther) errHandleFunc {
 	r := a.Middleware(getOrPOSTMiddleware(next))
 	r = http2Middleware(r)
 	return r
+}
+
+func sendAutoupdateData(w io.Writer, all bool, data map[string]json.RawMessage, fromChangeID, toChangeID int) error {
+	changed := make(map[string][]json.RawMessage)
+	deleted := make(map[string][]int)
+	for k := range data {
+		parts := strings.Split(k, ":")
+		if len(parts) != 2 {
+			return fmt.Errorf("invalid key %s, expected exacly one `:`", k)
+		}
+
+		collection := parts[0]
+		id, err := strconv.Atoi(parts[1])
+		if err != nil {
+			return fmt.Errorf("invalid key %s, id is not a number", k)
+		}
+
+		if data[k] == nil {
+			if !all {
+				deleted[collection] = append(deleted[collection], id)
+			}
+			continue
+		}
+
+		changed[collection] = append(changed[collection], data[k])
+	}
+
+	format := struct {
+		Changed      map[string][]json.RawMessage `json:"changed"`
+		Deleted      map[string][]int             `json:"deleted"`
+		FromChangeID int                          `json:"from_change_id"`
+		ToChangeID   int                          `json:"to_change_id"`
+		AllData      bool                         `json:"all_data"`
+	}{
+		changed,
+		deleted,
+		fromChangeID,
+		toChangeID,
+		all,
+	}
+
+	if err := json.NewEncoder(w).Encode(format); err != nil {
+		return fmt.Errorf("encode and send output data: %w", err)
+	}
+	w.(http.Flusher).Flush()
+	return nil
+}
+
+func projectorIDs(raw string) ([]int, error) {
+	parts := strings.Split(raw, ",")
+	ids := make([]int, len(parts))
+	for i, rpid := range parts {
+		id, err := strconv.Atoi(rpid)
+		if err != nil {
+			return nil, invalidRequestError{fmt.Errorf("projector_ids has to be a list of ints not `%s`", raw)}
+		}
+
+		ids[i] = id
+	}
+	return ids, nil
+}
+
+type connectionCount struct {
+	mu sync.Mutex
+	v  int
+}
+
+func (c *connectionCount) Add() int {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	c.v++
+	return c.v
+}
+
+func (c *connectionCount) Sub() int {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	c.v--
+	return c.v
 }
